Loading Mem locations asm-parsing
LDA,LDX,LDY now support loading mem locations zero page and 16 bit addresses.

+++ src/util.rs
@@ -14,8 +14,8 @@ pub fn is_zero_page(value: &str) -> bool {
         .unwrap_or_else(|_| panic!("Failed to parse hex value: {}", value));
     return converted_value < 256;
 }
-pub fn convert_string_to_u16(value: &str) -> u8 {
-    match value.parse::<u8>() {
+pub fn convert_string_to_u16(value: &str) -> u16 {
+    match value.parse::<u16>() {
         Ok(parsed_value) => return parsed_value,
         Err(_) => panic!("Failed to parse value as u8: {}", value),
     }
